Change auto-use algorithm.
When a ticket is scanned (or order number entered, or order chosen in
will call), the software will no longer propose using the quantity of
tickets ordered on each order line.  Instead it will propose using
exactly one ticket, of whatever ticket class appears in the first line
of the order.  This is more predictable behavior for the ticket taker.