Split APIs.
Rather than having one monolithic API, the CGI script now has separate
APIs for the office client, the payment form clients, and the point of
sale clients.  The clients themselves have been updated to use the new
APIs.  However, the old monolithic API is still accepted for the point
of sale calls, since older installations of the iOS app could be in use
for another 9 months or so.